+++ CHANGELOG.md
@@ -39,7 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Removed `compas.scene.SceneTree`, functionalities merged into `compas.scene.Scene`.
 * Removed default implementation of `compas.geometry.trimesh_geodistance` since nonexistent.
 * Removed `compas.utilities.geometric_key` and replaced it by `compas.tolerance.TOL.geometric_key`.
-* Removed `compas.utilities.geometric_key_xy`.
+* Removed `compas.utilities.geometric_key_xy` and replaced it by `compas.tolerance.TOL.geometric_key_xy`.
 
 ## [2.1.0] 2024-03-01